Update scaffold generator and templates for styling
Using the scaffold generator results in a lot of warnings for projects
using rubocop. This is mostly due to whitespace issues and using the
Ruby pre-1.9 hash syntax. This cleans up the output to avoid:

* tabs
* trailing whitespace
* inconsistent indentation
* extra line breaks
* Ruby pre-1.9 hash (hash rocket) syntax

If this gem is used in any applications using an older version of Ruby,
this change to the hash syntax will break compatibility. Due to this, the
minor version is being bumped.